+++ server/src/services/users.service.ts
@@ -1,11 +1,11 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, InternalServerErrorException } from '@nestjs/common';
 import { PrismaService } from '../services/prisma.service';
 import { createHash } from '../utils/hash';
 import { UserCreateDto } from '../dto/user-create.dto';
 
 @Injectable()
 export class UsersService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) { }
 
   async create(userCreate: UserCreateDto) {
     const userFound = await this.prisma.user.findUnique({
@@ -30,7 +30,14 @@ export class UsersService {
     return this.prisma.user.update({ where: { id }, data: { active: false } });
   }
   async delete(id: string) {
-    return this.prisma.user.delete({ where: { id } });
+    const userDeleted = await this.prisma.user.findUnique({ where: { id } });
+    if (!userDeleted) throw new BadRequestException('User Id inválido');
+    try {
+      await this.prisma.user.delete({ where: { id } });
+    } catch (error) {
+      throw new InternalServerErrorException(error)
+    }
+    return userDeleted
   }
   async findAll() {
     return this.prisma.user.findMany();
